pyproject.toml
[project]
name = "efficient_superpixel_segmentation"
version = "0.1.0"
description = "Efficient Superpixel Segmentation Project"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "scikit-image>=0.21.0",
    "Pillow>=9.5.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "tensorboard>=2.13.0",
    "opencv-python>=4.8.0",
    "einops>=0.6.1",
    "timm>=0.9.0",
    "torch-scatter>=2.1.0"
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
